pyproject.toml
[project]
name = "koroku"
version = "0.1.0"
description = "A Minimal Coding Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.1",
]
